requirements.txt
git+https://github.com/shahbuland/PerceptualSimilarity
torchmetrics
einops
ema-pytorch
git+https://github.com/KellerJordan/Muon
vector-quantize-pytorch
ray
omegaconf
wandb
kornia
datasets
huggingface-hub
